refactor(pipeline/api-direct-call): extract FlowTypes from IApiDirectCallConfig
Phase 8 commit 5/8 of ApiDirectCallConfig split.

- Carve FlowTypes (FlowKind, StepName, IStepConfig) from
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/IApiDirectCallConfig.ts.
- New file:
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/ConfigContracts/FlowTypes.ts
  (41 raw LoC). Imports IBodyTemplate / IEnvelopeSelectors /
  JsonValueTemplate from ./TemplateTypes.js, IPreStepHook from
  ./EnvelopeTypes.js, plus WKUrlGroup from ../../../Registry/WK —
  bucket-4 per the locked TemplateTypes/EnvelopeTypes -> FlowTypes
  chain.
- IApiDirectCallConfig.ts now reduced to 120 LoC: imports the three
  flow types and the WKUrlGroup direct import is dropped (now owned
  by FlowTypes). All 53 callers continue to compile via the existing
  `export type {}` re-export block.
- `npx madge --circular --extensions ts ConfigContracts/` reports
  zero cycles — the rubber-duck F4 acyclic chain holds end-to-end.

Verified:
- npm run type-check green
- npx tsup green (lib/index.cjs byte-identical to main)
- npx jest --testPathPatterns=ApiDirectCall green (223/223)
- npx eslint src/Scrapers/Pipeline/Mediator/ApiDirectCall green
- npx madge --circular ConfigContracts/ -> `No circular dependency found!`
- New sub-module 41 LoC (well under 150 cap)

Master plan: pipeline-decoupling-master-2026-05-28 / phase-8
