+++ src/Scrapers/Pipeline/Mediator/ApiDirectCall/IApiDirectCallConfig.ts
@@ -14,7 +14,6 @@
  * wiring time.
  */
 
-import type { WKUrlGroup } from '../../Registry/WK/UrlsWK.js';
 import type {
   IDerivedCarry,
   ISeedCarrySource,
@@ -28,6 +27,7 @@ import type {
   IPreStepHook,
   IProbeConfig,
 } from './ConfigContracts/EnvelopeTypes.js';
+import type { FlowKind, IStepConfig, StepName } from './ConfigContracts/FlowTypes.js';
 import type {
   AsymmetricSignerAlgorithm,
   CanonicalPart,
@@ -45,34 +45,6 @@ import type {
   JsonValueTemplate,
   RefToken,
 } from './ConfigContracts/TemplateTypes.js';
-
-/** Exhaustive flow-kind discriminator (spec.txt §B.3). */
-type FlowKind = 'sms-otp' | 'stored-jwt' | 'bearer-static';
-
-/** Step identifiers in the sms-otp flow. */
-type StepName = 'bind' | 'assertPassword' | 'assertOtp' | 'getIdToken' | 'sessionToken';
-
-/** Per-step config: name, URL tag, body template, response-extract selectors. */
-interface IStepConfig {
-  readonly name: StepName;
-  readonly urlTag: WKUrlGroup;
-  readonly body: IBodyTemplate;
-  readonly extractsToCarry: IEnvelopeSelectors;
-  /** Optional hook — awaited before the step fires. */
-  readonly preHook?: IPreStepHook;
-  /**
-   * Optional URL query params — hydrated JsonValueTemplate whose
-   * root is an object whose string-valued leaves become the
-   * outgoing ?k=v pairs. Values may be $ref / $literal / nested.
-   */
-  readonly queryTemplate?: JsonValueTemplate;
-  /**
-   * When true, captures this step's response Set-Cookie lines into
-   * the internal cookie jar. Subsequent steps with cookieJar=true
-   * include those cookies on the outbound Cookie header.
-   */
-  readonly cookieJar?: boolean;
-}
 
 /** Top-level config literal — placed into PIPELINE_BANK_CONFIG[bank].apiDirectCall. */
 interface IApiDirectCallConfig {
